MathChallenge 1-15-18
Line 107 keeps throwing an undefined error for the ToCharArray even though I've defined it and used that same type of loop twice earlier in the code. Everything else seems to be in working order except for that, and I need assistance with getting to the source of the error. I've reached out to some of my peers, Google, and StackOverflow but I must be missing something.